access_apis/requirements.txt
requests==2.25.0
pandas==1.1.4
numpy==1.19.4
slackclient==2.9.3
